+++ code/test/sixsq/nuvla/server/resources/deployment_log_lifecycle_test.clj
@@ -119,7 +119,7 @@
                           (ltu/is-status 200)
                           (ltu/is-operation-present :fetch)
                           (ltu/is-operation-present :edit)
-                          (ltu/is-operation-absent :delete))
+                          (ltu/is-operation-present :delete))
 
             fetch-url (ltu/get-op-url resp "fetch")
 
@@ -190,15 +190,8 @@
           (ltu/body->edn)
           (ltu/is-status 200))
 
-      ;; logs cannot be deleted by user
-      (-> session-jane
-          (request test-uri
-                   :request-method :delete)
-          (ltu/body->edn)
-          (ltu/is-status 403))
-
       ;; logs can be deleted by user
-      (-> session-admin
+      (-> session-jane
           (request test-uri
                    :request-method :delete)
           (ltu/body->edn)