+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Change
+
+  - Deployment log - resource can be deleted by user session
+
 ## [4.2.3] - 2019-09-04
 
 ### Added